Double is replaced by int

+++ TestCalculator.java
@@ -1,4 +1,3 @@
-
 import java.util.InputMismatchException;
 import java.util.Scanner;
 
@@ -12,34 +11,34 @@ public class TestCalculator {
         System.out.println("Press 4 to divide numbers");
         System.out.println("Press 0 to exit");
             try {
-                int dec = scanner.nextInt();
-                switch (dec) {
+                double dec = scanner.nextInt();
+                switch ((int) dec) {
                     case 1 -> {
                         System.out.println("Enter first number");
-                        calculate.setNumber1(scanner.nextInt());
+                        calculate.setNumber1(scanner.nextDouble());
                         System.out.println("Enter second number");
-                        calculate.setNumber2(scanner.nextInt());
+                        calculate.setNumber2(scanner.nextDouble());
                         System.out.println("Sum of numbers : " + Calculator.add(calculate.getNumber1(), calculate.getNumber2()));
                     }
                     case 2 -> {
                         System.out.println("Enter first number");
-                        calculate.setNumber1(scanner.nextInt());
+                        calculate.setNumber1(scanner.nextDouble());
                         System.out.println("Enter second number");
-                        calculate.setNumber2(scanner.nextInt());
+                        calculate.setNumber2(scanner.nextDouble());
                         System.out.println("Subtraction of numbers : " + Calculator.subtract(calculate.getNumber1(), calculate.getNumber2()));
                     }
                     case 3 -> {
                         System.out.println("Enter first number");
-                        calculate.setNumber1(scanner.nextInt());
+                        calculate.setNumber1(scanner.nextDouble());
                         System.out.println("Enter second number");
-                        calculate.setNumber2(scanner.nextInt());
+                        calculate.setNumber2(scanner.nextDouble());
                         System.out.println("Product of numbers : " + Calculator.product(calculate.getNumber1(), calculate.getNumber2()));
                     }
                     case 4 -> {
                         System.out.println("Enter first number");
-                        calculate.setNumber1(scanner.nextInt());
+                        calculate.setNumber1(scanner.nextDouble());
                         System.out.println("Enter second number");
-                        calculate.setNumber2(scanner.nextInt());
+                        calculate.setNumber2(scanner.nextDouble());
                         System.out.println("Division of numbers is : " + Calculator.division(calculate.getNumber1(), calculate.getNumber2()));
 
                     }
